Remove timeout manager
