--- src/NServiceBus.Core/DelayedDelivery/TimeoutManagerRoutingStrategy.cs
@@ -1,28 +1,0 @@
-namespace NServiceBus
-{
-    using System;
-    using System.Collections.Generic;
-    using Routing;
-
-    class TimeoutManagerRoutingStrategy : RoutingStrategy
-    {
-        string timeoutManagerAddress;
-        string ultimateDestination;
-        DateTimeOffset deliverAt;
-
-        public TimeoutManagerRoutingStrategy(string timeoutManagerAddress, string ultimateDestination, DateTimeOffset deliverAt)
-        {
-            this.ultimateDestination = ultimateDestination;
-            this.deliverAt = deliverAt;
-            this.timeoutManagerAddress = timeoutManagerAddress;
-        }
-
-        public override AddressTag Apply(Dictionary<string, string> headers)
-        {
-            headers[TimeoutManagerHeaders.RouteExpiredTimeoutTo] = ultimateDestination;
-            headers[TimeoutManagerHeaders.Expire] = DateTimeOffsetHelper.ToWireFormattedString(deliverAt);
-
-            return new UnicastAddressTag(timeoutManagerAddress);
-        }
-    }
-}
